% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parmsurvfit.R
\docType{data}
\name{aggressive}
\alias{aggressive}
\title{Data on time until drivers honked their horn when being blocked from an intersection}
\format{A data frame with 57 rows and 2 variables:
\describe{
  \item{seconds}{Number of seconds until showing signs of aggression}
  \item{censor}{censoring status indicator variable (0 = censored event time, 1 = complete event time)}
}}
\source{
{https://stats.idre.ucla.edu/other/examples/alda/}
}
\usage{
aggressive
}
\description{
Diekmann et al. (1996) investigated the association between driver
characteristics and social status of cars to aggressive driver
responses by measuring the time that elapsed between the being
blocked and honking the horn. Researchers intentionally blocked 
57 motorists at a green light by a Volkswagen Jetta, and recorded 
the time it took for motorists to show signs of aggression.
Signs of aggression included honking their horn or beaming 
the headlights at the Jetta
}
\keyword{datasets}
