% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graduate.R
\docType{data}
\name{graduate}
\alias{graduate}
\title{Data on time until graduation for 1000 college students.}
\format{A data frame with 1000 rows and 3 variables:
\describe{
  \item{Years}{years until graduation}
  \item{Censor}{censoring status indicator variable (0 = censored event time)}
  \item{Gender}{a dichotomous variable identifying gender (male = 0, female = 1)}
}}
\source{
\url{}
}
\usage{
graduate
}
\description{
A dataset that contains the time (in years) that 1000 students (472 males and
528 females) took to graduate (obtain a bachelor’s degree) from college (measured from the time
they entered a post-secondary institution, i.e. either a junior college or four year degree granting
institution). The Gender column contains the gender of each student (0 = male, 1 = female), and
Censor contains the values of the censoring status variable.
}
\keyword{datasets}
