% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parmsurvfit.R
\docType{data}
\name{rearrest}
\alias{rearrest}
\title{Data on time until reincarceration for 194 inmates.}
\format{A data frame with 194 rows and 5 variables:
\describe{
  \item{Months}{months until re-arrest}
  \item{Censor}{censoring status indicator variable (0 = censored event time, 1 = complete event time)}
  \item{Personal}{a dichotomous variable identifying former inmates who had a history of
  person-related crimes (personal = 1), i.e. those with one or more convictions for offenses
  such as aggravated assault or kidnapping}
  \item{Property}{a dichotomous variable indicating whether former inmates were convicted
  of a property-related crime (property = 1)}
  \item{Cenage}{the "centered" age of individual, i.e. the difference between the age of the
  individual upon release and the average age of all inmates in the study.}
}}
\source{
\url{http://www.ats.ucla.edu/stat/examples/alda.htm}
}
\usage{
rearrest
}
\description{
Henning and Frueh (1996) followed criminal activities of 194 inmates released from a medium
security prison for 36 months. The data from this study can be used to investigate the time
until the former inmates were re-arrested. If the former inmate had been re-arrested for
a criminal act before 36 months (after initial prison release) had passed, then that former
inmate’s event time was complete. If the former inmate had not been re-arrested for a
criminal act after 36 months had passed, or had completely dropped out of the study, then
that former inmate’s event time was right censored.
}
\keyword{datasets}
